+++ protected/humhub/components/Module.php
@@ -8,10 +8,14 @@
 
 namespace humhub\components;
 
+use humhub\modules\activity\components\BaseActivity;
+use humhub\modules\activity\models\Activity;
 use humhub\modules\file\libs\FileHelper;
 use humhub\modules\notification\components\BaseNotification;
 use Yii;
 use yii\helpers\Json;
+use yii\web\AssetBundle;
+use yii\web\HttpException;
 
 /**
  * Base Class for Modules / Extensions
@@ -161,10 +165,22 @@ class Module extends \yii\base\Module
 
     /**
      * Publishes the basePath/resourcesPath (assets) module directory if existing.
+     * @param bool $all whether or not to publish sub assets within the `assets` directory
      * @return array
      */
-    public function publishAssets()
+    public function publishAssets($all = false)
     {
+        /** @var $assetBundle AssetBundle */
+        /** @var $manager AssetManager */
+
+        if($all) {
+            foreach ($this->getAssetClasses() as $assetClass) {
+                $assetBundle = new $assetClass();
+                $manager = Yii::$app->getAssetManager();
+                $manager->forcePublish($assetBundle);
+            }
+        }
+
         if ($this->hasAssets()) {
             return Yii::$app->assetManager->publish($this->getAssetPath(), ['forceCopy' => true]);
         }
@@ -386,11 +402,43 @@ class Module extends \yii\base\Module
         $activityDirectory = $this->getBasePath() . DIRECTORY_SEPARATOR . 'activities';
         if (is_dir($activityDirectory)) {
             foreach (FileHelper::findFiles($activityDirectory, ['recursive' => false,]) as $file) {
-                $activities[] = $activityNamespace . '\\' . basename($file, '.php');
+                $activityClass = $activityNamespace . '\\' . basename($file, '.php');
+                if(is_subclass_of($activityClass, BaseActivity::class)) {
+                    $activities[] = $activityClass;
+                }
             }
         }
 
         return $activities;
+    }
+
+    /**
+     * Returns a list of asset class names this modules provides.
+     *
+     * @since 1.2.8
+     * @return array list of asset class names
+     */
+    public function getAssetClasses()
+    {
+        $class = get_class($this);
+        if (($pos = strrpos($class, '\\')) !== false) {
+            $assetNamespace = substr($class, 0, $pos) . '\\assets';
+        } else {
+            $assetNamespace = '';
+        }
+
+        $assets = [];
+        $assetDirectory = $this->getBasePath() . DIRECTORY_SEPARATOR . 'assets';
+        if (is_dir($assetDirectory)) {
+            foreach (FileHelper::findFiles($assetDirectory, ['recursive' => false,]) as $file) {
+                $assetClass =  $assetNamespace . '\\' . basename($file, '.php');
+                if(is_subclass_of($assetClass, AssetBundle::class)) {
+                    $assets[] = $assetClass;
+                }
+            }
+        }
+
+        return $assets;
     }
 
 }